+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Next Version
+
+# Version 4.3.1
 - Artifacts are now delivered through `Maven Central` repository instead of `JCenter` which is deprecated.
 - Handle retro compatibility of AdMob < v19.7.0